changed the response to Json for circle and mailbox api

+++ dao/cassandra/mailbox/index.js
@@ -23,12 +23,16 @@ function checkIfMailboxExists(mailboxId, callback) {
 }
 // Function to delete the mailbox with id. If id not exists returns no mailbox error
 function deleteMailbox(mailboxId, callback) {
-  const query = (`DELETE from  mailbox where id =${mailboxId}`);
-  client.execute(query, (err, result) => {
+  checkIfMailboxExists(mailboxId, (err, mailboxExists) => {
     if (err) { return callback(err, null); }
-    return callback(err, mailboxId);
+    if (mailboxExists===false) {
+      return callback(`circle id ${mailboxId} does not exis`, null);
+    } else {
+      const query = (`DELETE from  mailbox where id =${mailboxId}`);
+      client.execute(query, (error, result) => callback(error, mailboxId));
+      return { mailboxId };
+    }
   });
-  return true;
 }
